Split ministep out of alongstep test
