+++ test/CMakeLists.txt
@@ -231,9 +231,9 @@ add_library(testcel_celeritas
   celeritas/GeantTestBase.cc
   celeritas/GlobalTestBase.cc
   celeritas/GlobalGeoTestBase.cc
+  celeritas/MiniStepTestBase.cc
   celeritas/MockTestBase.cc
   celeritas/SimpleTestBase.cc
-  celeritas/global/AlongStepTestBase.cc
   celeritas/global/StepperTestBase.cc
 )
 celeritas_target_link_libraries(testcel_celeritas
